fix: split client-based tests into separate file (CI event loop compat)
Python 3.12 CI fails when db_session-only tests and client-based tests
share a file — the session-scoped engine fixture and client's create_app
engine conflict during teardown. Splitting into separate files isolates
the event loop lifecycles.
